fix remarks in fiveth task
