+++ src/Task3DelegatesEventsExceptions/INPUT.txt
@@ -1,5 +1,5 @@
-B 0 9 7
-Z 1 2 3 4
-A 23 48 44
-Z 36 92 101
-T -2 -6 0 12
+U 0 9 7
+T 1 2 3 4
+K 23 48 44
+C 36 92 101
+Z -2 -6 0 12
